+++ index.html
@@ -8,6 +8,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
+  <script src="js/main.js" type="module"></script>
 </head>
 
 <body>
@@ -241,7 +242,5 @@
       </div>
     </section>
   </template>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
 </body>
 </html>
